feat: Auto-upgrade multiple levels
In some situations such as a new unit being unlocked later in the game (like the tacjet) when the player already has a high affinity level, multiple unit upgrades can be available at once. Previously the unit would only auto-upgrade the first level, then the next turn it would auto-upgrade the next level. But if there are other unit upgrades at the same time it can be a bit confusing in terms of UX.

Closes https://github.com/bmaupin/micro-beyond-earth/issues/11